+++ spec/page_feedback/engine_spec.rb
@@ -5,7 +5,8 @@ require "rails_helper"
 RSpec.describe PageFeedback::Engine do
   let(:browser_modules) do
     %w[
-      page_feedback/capture_controller_support page_feedback/clipboard page_feedback/context_recorder
+      page_feedback/capture_controller_support page_feedback/capture_mode_status
+      page_feedback/clipboard page_feedback/context_recorder
       page_feedback/controllers/capture_controller page_feedback/controllers/copy_controller
       page_feedback/element_capture
       page_feedback/feedback_picker page_feedback/review_highlight page_feedback/visual_viewport